Working HTTP server from service

+++ ServiceHosts/MPExtended.ServiceHosts.WebMediaPortal/HostService.cs
@@ -24,7 +24,6 @@ using System.Linq;
 using System.ServiceProcess;
 using System.Text;
 using System.IO;
-using System.Diagnostics;
 using System.Threading;
 using MPExtended.Libraries.General;
 
@@ -49,14 +48,14 @@ namespace MPExtended.ServiceHosts.WebMediaPortal
 #endif
 
             string iisExpress = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), "IIS Express", "iisexpress.exe");
-            string arguments = String.Format("/path:{0} /port:{1}", path, 8080);
+            string arguments = String.Format("/systray:0 /path:{0} /port:{1}", path, 8080);
             string logPath = Path.Combine(Installation.GetLogDirectory(), String.Format("WebMediaPortalIIS-{0:yyyy_MM_dd}.log", DateTime.Now));
 
             hostProcess = new Process();
             hostProcess.StartInfo = new ProcessStartInfo()
             {
                 FileName = iisExpress,
-                Arguments = "",
+                Arguments = arguments,
                 RedirectStandardOutput = true,
                 UseShellExecute = false,
                 CreateNoWindow = true
@@ -65,16 +64,17 @@ namespace MPExtended.ServiceHosts.WebMediaPortal
 
             logThread = new Thread(delegate(object param)
             {
-                Stream data = (Stream)param;
-                using (StreamReader reader = new StreamReader(data))
+                using (StreamReader reader = (StreamReader)param)
                 {
-                    using (StreamWriter writer = new StreamWriter(logPath, true))
+                    using (StreamWriter writer = new StreamWriter(logPath, true, Encoding.UTF8, 16 * 1024))
                     {
+                        writer.WriteLine("<process started at {0:yyyy-MM-dd HH:mm:ss} with arguments {1}>", DateTime.Now, arguments);
                         string line;
                         while ((line = reader.ReadLine()) != null)
                         {
-                            writer.WriteLine(line);
+                            writer.WriteLine("[{0:yyyy-MM-dd HH:mm:ss}] {1}", DateTime.Now, line);
                         }
+                        writer.WriteLine("<process exited at {0:yyyy-MM-dd HH:mm:ss}>", DateTime.Now);
                     }
                 }
 
@@ -84,8 +84,26 @@ namespace MPExtended.ServiceHosts.WebMediaPortal
 
         protected override void OnStop()
         {
-            hostProcess.Kill();
-            logThread.Join();
+            try
+            {
+                if(!hostProcess.HasExited)
+                    hostProcess.Kill();
+            }
+            catch (Exception)
+            {
+                // we can't do much about it with our logging etc already shutdown
+            }
+
+            try
+            {
+                logThread.Join(TimeSpan.FromSeconds(5));
+                if (logThread.IsAlive)
+                    logThread.Abort();
+            }
+            catch (Exception)
+            {
+                // we can't do much about it with our logging etc already shutdown
+            }
         }
     }
 }
